+++ example/lib/gallery/sections/sections.dart
@@ -11,6 +11,7 @@ export 'key_icons.dart';
 export 'screen_transitions.dart';
 export 'selection_list.dart';
 export 'single_child_scroll_view.dart';
+export 'split_panel.dart';
 export 'tab_view.dart';
 export 'text_fields.dart';
 export 'texts.dart';