+++ example/lib/gallery/sections/sections.dart
@@ -3,6 +3,7 @@ export 'checkboxes.dart';
 export 'containers.dart';
 export 'custom_extensions.dart';
 export 'dialogs.dart';
+export 'drop_shadows.dart';
 export 'file_explorer.dart';
 export 'icon_buttons.dart';
 export 'icons.dart';