+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # 0.6.0
 
  - feat: add `NesIcons.threeVerticalDots`
+ - feat: add `NesIcons.threeHorizontalDots`
  - fix: `NesScrollbar` has the wrong border colors
  - feat: add `NesSingleChildScrollview.clipContent`
  - feat: add `NesTabView`
  - feat: Update `NesButton` widget to enable/disable based on onPressed value
+ - feat: add `NesSplitPanel`
 
 # 0.5.0